+++ app/src/main/java/com/weatherapp/models/Main.kt
@@ -3,9 +3,9 @@ package com.weatherapp.models
 import java.io.Serializable
 
 data class Main(
-    val temp: Int,
-    val pressure: Int,
-    val humidity: Int,
-    val tempMin: Int,
-    val tempMax: Int
+    val temp: Double,
+    val pressure: Double,
+    val humidity: Double,
+    val tempMin: Double,
+    val tempMax: Double
 ) : Serializable
