+++ app/src/main/java/com/weatherapp/models/Sys.kt
@@ -3,9 +3,9 @@ package com.weatherapp.models
 import java.io.Serializable
 
 data class Sys(
-    val type: Int,
+    val type: Double,
     val message: Double,
     val country: String,
-    val sunrise: Int,
-    val sunset: Int
+    val sunrise: Double,
+    val sunset: Double
 ) : Serializable
