+++ app/src/main/java/com/weatherapp/models/Weather.kt
@@ -3,7 +3,7 @@ package com.weatherapp.models
 import java.io.Serializable
 
 data class Weather(
-    val id: Int,
+    val id: Double,
     val main: String,
     val description: String,
     val icon: String
