+++ app/src/main/java/com/weatherapp/models/Wind.kt
@@ -4,5 +4,5 @@ import java.io.Serializable
 
 data class Wind(
     val speed: Double,
-    val deg: Int
+    val deg: Double
 ) : Serializable
